Move reusable random generators to btrie_test.go.
